requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
fastapi==0.115.0
h11==0.16.0
httptools==0.6.4
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.38.6
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1

# Database stack
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13

# R2 / uploads
boto3>=1.35
python-multipart>=0.0.9

# Background jobs
rq>=2.6.0
redis>=7.0.1